did something to GenericImageDrawer

+++ src/lisp/drawers/GenericImageDrawer.java
@@ -8,11 +8,12 @@ import java.io.IOException;
 import javax.imageio.ImageIO;
 
 import lisp.GameObject;
+import lisp.WithImage;
 
 public abstract class GenericImageDrawer<T extends WithImage & GameObject> implements GameObjectDrawer<T> {
 
 	@Override
 	public void draw(T go, Graphics2D g) {
-		g.drawImage(go.getImage(), go.getX(), go.getY(), go.getWidth(), go.getHeight(), null);
+		g.drawImage(go.getImage(), (int)go.getX(), (int)go.getY(), (int)go.getWidth(), (int)go.getHeight(), null);
 	}
 }
